class Usersalong with variables are added
